genetic linkage works, fix test #30

+++ DispersalTrait.h
@@ -57,7 +57,7 @@ public:
 		int startChr) {
 		(this->*_inherit_func_ptr)(fromMother, parentGenes, recomPositions, startChr);
 	}
-	float getAlleleIDAtLocus(short whichChromosome, int position) const;
+	int getAlleleIDAtLocus(short whichChromosome, int position) const;
 #endif
 
 private:
